One more fix

+++ variants/eval_pooled2.sh
@@ -104,6 +104,10 @@ mkdir -p ${output}
 
 for cn in "${cns[@]}"; do
     subdir="${output}/${cn}"
+    if [[ -f "${subdir}/eval/done" ]] && [[ -f "${subdir}/eval_squash/done" ]]; then
+        echo "Skipping ${subdir}"
+        continue
+    fi
     rm -rf "${subdir}"
     mkdir -p "${subdir}"
 
@@ -130,34 +134,26 @@ for cn in "${cns[@]}"; do
     fi
     echo -e "\n=== Examining ${reg_stats[0]} regions, in total ${reg_stats[1]} bp ==="
 
-    if [[ -f "${subdir}/eval/done" ]]; then
-        echo "Skipping evaluation"
-    else
-        rtg "RTG_MEM=${mem}" vcfeval \
-            -e "${subdir}/eval.bed" \
-            --decompose \
-            -b "${subdir}/bench.vcf.gz" \
-            -c "${subdir}/calls.vcf.gz" \
-            -t "${genome}.sdf" \
-            -o "${subdir}/eval" \
-            --sample-ploidy=${cn} &> "${subdir}/eval.log"
-    fi
+    rtg "RTG_MEM=${mem}" vcfeval \
+        -e "${subdir}/eval.bed" \
+        --decompose \
+        -b "${subdir}/bench.vcf.gz" \
+        -c "${subdir}/calls.vcf.gz" \
+        -t "${genome}.sdf" \
+        -o "${subdir}/eval" \
+        --sample-ploidy=${cn} &> "${subdir}/eval.log"
     ${wdir}/write_summary.py "${subdir}/eval"
 
     echo -e "\n=== Evaluating squashed variants ==="
-    if [[ -f "${subdir}/eval_squash/done" ]]; then
-        echo "Skipping evaluation"
-    else
-        rtg "RTG_MEM=${mem}" vcfeval \
-            -e "${subdir}/eval.bed" \
-            --decompose \
-            -b "${subdir}/bench.vcf.gz" \
-            -c "${subdir}/calls.vcf.gz" \
-            -t "${genome}.sdf" \
-            -o "${subdir}/eval_squash" \
-            --squash-ploidy \
-            --sample-ploidy=${cn} &> "${subdir}/eval_squash.log"
-    fi
+    rtg "RTG_MEM=${mem}" vcfeval \
+        -e "${subdir}/eval.bed" \
+        --decompose \
+        -b "${subdir}/bench.vcf.gz" \
+        -c "${subdir}/calls.vcf.gz" \
+        -t "${genome}.sdf" \
+        -o "${subdir}/eval_squash" \
+        --squash-ploidy \
+        --sample-ploidy=${cn} &> "${subdir}/eval_squash.log"
     ${wdir}/write_summary.py "${subdir}/eval_squash"
     echo
 done
